+++ gluecodium/src/test/resources/smoke/external_types/output/android/com/example/smoke/UseJavaExternalConst.java
@@ -4,7 +4,7 @@
 package com.example.smoke;
 import android.support.annotation.NonNull;
 public final class UseJavaExternalConst {
-    public static final com.example.Foo DEFAULT_TRUTH = new com.example.Foo("foo");
+    static final com.example.Foo DEFAULT_TRUTH = new com.example.Foo("foo");
     @NonNull
     public String stringField;
     public UseJavaExternalConst(@NonNull final String stringField) {